Add files via upload
I reworked prepocessesCTX.sh to be more up to date and to map project the files into the correct schema using gdal. Now I can run this script and not need to access the USGS astrogeology website to preprocess the images for me. The CTXDTMgen is a spinoff of preprocessCTX and will create a CTX DTM based off of the two images that exist within the local directory.